% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_shinytitle_example}
\alias{run_shinytitle_example}
\title{Run shinytitle Example}
\usage{
run_shinytitle_example(example = c("toggle", "busy"), ...)
}
\arguments{
\item{example}{Choose between \code{toggle} and \code{busy}}

\item{...}{other arguments sent to \code{\link[shiny]{runApp}}}
}
\description{
Launch one of the examples contained in the \code{shinytitle} package:

\itemize{
\item{toggle}{An example showing the effects of simple change to the title and flashing title}
\item{busy}{An example of when the title changes when the shiny app is busy running calculations}
}
}
\examples{
if (interactive()) {
  library(shiny)
  run_shinytitle_example()
}

}
