% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_shiny_title}
\alias{use_shiny_title}
\title{Enable \code{shinytitle}}
\usage{
use_shiny_title()
}
\description{
Adding script to enable the ability to change the window title.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    title = "Initial Title",
    use_shiny_title(),
    actionButton("button", "Click me for a new title"),
  )

  server <- function(input, output, session) {
    observeEvent(input$button, {
      change_window_title(session, "New Title")
    })
  }

  shinyApp(ui, server)
}

}
\seealso{
\code{\link{change_window_title}} \code{\link{flashing_window_title}}
}
